feature/chgres_noahmp  This commit references issue #3.
Update surface.F90 as follows:

When using LIS, there were a few points with mismatches between the number of
snow layers (snowxy) and the layered snow depth (zsnsoxy).  Don't know if this
is the result of bad LIS data or how chgres is interpolating these fields.  It does not
happen with all cases.  But mismatches will cause the forecast model to break.
For now, set a check for mismatches.  If one exists, remove all snow at
the mismatched point.